Refactor: [BADA-227] 예약 API 관련 코드 분리 및 리팩토링 #142
