--- src/pages/rental/store/reservation/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as DeviceReceiptItem } from './DeviceReceiptItem';
